Rename directory and namespace to "system" from "core"
