--- include/meevax/core/string.hpp
@@ -1,16 +1,0 @@
-#ifndef INCLUDED_MEEVAX_CORE_STRING_HPP
-#define INCLUDED_MEEVAX_CORE_STRING_HPP
-
-#include <meevax/character/unicode.hpp>
-#include <meevax/core/pair.hpp>
-
-namespace meevax::core
-{
-  struct string
-    : public virtual pair
-  {
-  };
-} // namespace meevax::core
-
-#endif // INCLUDED_MEEVAX_CORE_STRING_HPP
-
