+++ include/meevax/system/character.hpp
@@ -1,11 +1,11 @@
-#ifndef INCLUDED_MEEVAX_CORE_CHARACTER_HPP
-#define INCLUDED_MEEVAX_CORE_CHARACTER_HPP
+#ifndef INCLUDED_MEEVAX_SYSTEM_CHARACTER_HPP
+#define INCLUDED_MEEVAX_SYSTEM_CHARACTER_HPP
 
 #include <utility>
 
 #include <meevax/character/unicode.hpp>
 
-namespace meevax::core
+namespace meevax::system
 {
   struct character
     : public character::unicode<8>
@@ -15,7 +15,7 @@ namespace meevax::core
       : character::unicode<8> {std::forward<Ts>(args)...}
     {}
   };
-} // namespace meevax::core
+} // namespace meevax::system
 
-#endif // INCLUDED_MEEVAX_CORE_CHARACTER_HPP
+#endif // INCLUDED_MEEVAX_SYSTEM_CHARACTER_HPP
