+++ include/meevax/system/string.hpp
@@ -1,0 +1,16 @@
+#ifndef INCLUDED_MEEVAX_SYSTEM_STRING_HPP
+#define INCLUDED_MEEVAX_SYSTEM_STRING_HPP
+
+#include <meevax/character/unicode.hpp>
+#include <meevax/system/pair.hpp>
+
+namespace meevax::system
+{
+  struct string
+    : public virtual pair
+  {
+  };
+} // namespace meevax::system
+
+#endif // INCLUDED_MEEVAX_SYSTEM_STRING_HPP
+
